+++ parsers/dairy_disaster_parser.py
@@ -151,11 +151,11 @@ class DairyDisasterParser:
             new_data_entry = {
                 "state": state_abbr,
                 "subtitleName": "Dairy Margin Coverage, Subtitle D",
-                "totalCounts": int(sum_by_dairy_count_by_state[state].item()),
+                "totalRecipientCount": int(sum_by_dairy_count_by_state[state].item()),
                 "totalPaymentInDollars": round(sum_by_dairy_payments_by_state[state].item(), 2),
                 "averageRecipientCount": round(sum_by_dairy_count_by_state[state].item() / total_years, 2),
                 "totalPaymentInPercentageNationwide": dairy_payments_percentage_nation,
-                "totalCountsInPercentageNationwide": dairy_count_percentage_nation,
+                "totalRecipientCountInPercentageNationwide": dairy_count_percentage_nation,
                 "averageRecipientCountInPercentageNationwide": average_dairy_count_percentage_nation,
                 "programs": []
             }
@@ -185,7 +185,7 @@ class DairyDisasterParser:
         self.dairy_program_data_dict = {
             "subtitleName": "Dairy Margin Coverage, Subtitle D",
             "totalPaymentInDollars": round(dairy_total_by_payments.item(), 2),
-            "totalCounts": int(dairy_total_by_count.item()),
+            "totalRecipientCount": int(dairy_total_by_count.item()),
             "averageRecipientCount": round(dairy_total_by_count.item() / total_years, 2),
             "programs": []
         }
@@ -435,66 +435,66 @@ class DairyDisasterParser:
             new_data_entry = {
                 "state": state_abbr,
                 "subtitleName": "Supplemental Agricultural Disaster Assistance, Subtitle E",
-                "totalCounts": int(sum_by_disaster_count_by_state[state].item()),
+                "totalRecipientCount": int(sum_by_disaster_count_by_state[state].item()),
                 "totalPaymentInDollars": round(sum_by_disaster_payments_by_state[state].item(), 2),
                 "averageRecipientCount": round(sum_by_disaster_count_by_state[state].item() / total_years, 2),
                 "totalPaymentInPercentageNationwide": disaster_payments_percentage_nation,
-                "totalCountsInPercentageNationwide": disaster_count_percentage_nation,
+                "totalRecipientCountInPercentageNationwide": disaster_count_percentage_nation,
                 "averageRecipientCountInPercentageNationwide": average_disaster_count_percentage_nation,
                 "programs": [
                     {
                         "programName": "Emergency Assistance for Livestock, Honey Bees, and Farm-Raised Fish Program (ELAP)",
-                        "totalCounts": int(sum_by_elap_count_by_state[state].item()),
+                        "totalRecipientCount": int(sum_by_elap_count_by_state[state].item()),
                         "totalPaymentInDollars": round(sum_by_elap_payments_by_state[state].item(), 2),
                         "averageRecipientCount": round(sum_by_elap_count_by_state[state].item() / total_years,
                                                        2),
                         "totalPaymentInPercentageNationwide": elap_payments_percentage_nation,
-                        "totalCountsInPercentageNationwide": elap_count_percentage_nation,
+                        "totalRecipientCountInPercentageNationwide": elap_count_percentage_nation,
                         "averageRecipientCountInPercentageNationwide": average_elap_count_percentage_nation,
                         "totalPaymentInPercentageWithinState": within_state_elap_percentage_payments,
-                        "totalCountsInPercentageWithinState": within_state_elap_percentage_count,
+                        "totalRecipientCountInPercentageWithinState": within_state_elap_percentage_count,
                         "averageRecipientCountInPercentageWithinState": within_state_elap_average_percentage_count,
                         "subPrograms": []
                     },
                     {
                         "programName": "Livestock Forage Program (LFP)",
-                        "totalCounts": int(sum_by_lfp_count_by_state[state].item()),
+                        "totalRecipientCount": int(sum_by_lfp_count_by_state[state].item()),
                         "totalPaymentInDollars": round(sum_by_lfp_payments_by_state[state].item(), 2),
                         "averageRecipientCount": round(sum_by_lfp_count_by_state[state].item() / total_years,
                                                        2),
                         "totalPaymentInPercentageNationwide": lfp_payments_percentage_nation,
-                        "totalCountsInPercentageNationwide": lfp_count_percentage_nation,
+                        "totalRecipientCountInPercentageNationwide": lfp_count_percentage_nation,
                         "averageRecipientCountInPercentageNationwide": average_lfp_count_percentage_nation,
                         "totalPaymentInPercentageWithinState": within_state_lfp_percentage_payments,
-                        "totalCountsInPercentageWithinState": within_state_lfp_percentage_count,
+                        "totalRecipientCountInPercentageWithinState": within_state_lfp_percentage_count,
                         "averageRecipientCountInPercentageWithinState": within_state_lfp_average_percentage_count,
                         "subPrograms": []
                     },
                     {
                         "programName": "Livestock Indemnity Payments (LIP)",
-                        "totalCounts": int(sum_by_lip_count_by_state[state].item()),
+                        "totalRecipientCount": int(sum_by_lip_count_by_state[state].item()),
                         "totalPaymentInDollars": round(sum_by_lip_payments_by_state[state].item(), 2),
                         "averageRecipientCount": round(sum_by_lip_count_by_state[state].item() / total_years,
                                                        2),
                         "totalPaymentInPercentageNationwide": lip_payments_percentage_nation,
-                        "totalCountsInPercentageNationwide": lip_count_percentage_nation,
+                        "totalRecipientCountInPercentageNationwide": lip_count_percentage_nation,
                         "averageRecipientCountInPercentageNationwide": average_lip_count_percentage_nation,
                         "totalPaymentInPercentageWithinState": within_state_lip_percentage_payments,
-                        "totalCountsInPercentageWithinState": within_state_lip_percentage_count,
+                        "totalRecipientCountInPercentageWithinState": within_state_lip_percentage_count,
                         "averageRecipientCountInPercentageWithinState": within_state_lip_average_percentage_count,
                         "subPrograms": []
                     },
                     {
                         "programName": "Tree Assistance Program (TAP)",
-                        "totalCounts": int(sum_by_tap_count_by_state[state].item()),
+                        "totalRecipientCount": int(sum_by_tap_count_by_state[state].item()),
                         "totalPaymentInDollars": round(sum_by_tap_payments_by_state[state].item(), 2),
                         "averageRecipientCount": round(sum_by_tap_count_by_state[state].item() / total_years,
                                                        2),
                         "totalPaymentInPercentageNationwide": tap_payments_percentage_nation,
-                        "totalCountsInPercentageNationwide": tap_count_percentage_nation,
+                        "totalRecipientCountInPercentageNationwide": tap_count_percentage_nation,
                         "averageRecipientCountInPercentageNationwide": average_tap_count_percentage_nation,
                         "totalPaymentInPercentageWithinState": within_state_tap_percentage_payments,
-                        "totalCountsInPercentageWithinState": within_state_tap_percentage_count,
+                        "totalRecipientCountInPercentageWithinState": within_state_tap_percentage_count,
                         "averageRecipientCountInPercentageWithinState": within_state_tap_average_percentage_count,
                         "subPrograms": []
                     }
@@ -541,31 +541,31 @@ class DairyDisasterParser:
         self.disaster_program_data_dict = {
             "subtitleName": "Supplemental Agricultural Disaster Assistance, Subtitle E",
             "totalPaymentInDollars": round(disaster_total_by_payments.item(), 2),
-            "totalCounts": int(disaster_total_by_count.item()),
+            "totalRecipientCount": int(disaster_total_by_count.item()),
             "averageRecipientCount": round(disaster_total_by_count.item() / total_years, 2),
             "programs": [
                 {
                     "programName": "Emergency Assistance for Livestock, Honey Bees, and Farm-Raised Fish Program (ELAP)",
                     "totalPaymentInDollars": round(elap_total_by_payments.item(), 2),
-                    "totalCounts": int(elap_total_by_count.item()),
+                    "totalRecipientCount": int(elap_total_by_count.item()),
                     "averageRecipientCount": round(elap_total_by_count.item() / total_years, 2),
                 },
                 {
                     "programName": "Livestock Forage Program (LFP)",
                     "totalPaymentInDollars": round(lfp_total_by_payments.item(), 2),
-                    "totalCounts": int(lfp_total_by_count.item()),
+                    "totalRecipientCount": int(lfp_total_by_count.item()),
                     "averageRecipientCount": round(lfp_total_by_count.item() / total_years, 2),
                 },
                 {
                     "programName": "Livestock Indemnity Payments (LIP)",
                     "totalPaymentInDollars": round(lip_total_by_payments.item(), 2),
-                    "totalCounts": int(lip_total_by_count.item()),
+                    "totalRecipientCount": int(lip_total_by_count.item()),
                     "averageRecipientCount": round(lip_total_by_count.item() / total_years, 2),
                 },
                 {
                     "programName": "Tree Assistance Program (TAP)",
                     "totalPaymentInDollars": round(tap_total_by_payments.item(), 2),
-                    "totalCounts": int(tap_total_by_count.item()),
+                    "totalRecipientCount": int(tap_total_by_count.item()),
                     "averageRecipientCount": round(tap_total_by_count.item() / total_years, 2),
                 }
             ]